requirements.txt
asgiref==3.8.1
attrs==20.3.0
barcode==1.0.4
bytes==0.0.1.dev4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-cors-headers==4.6.0
django-default-imagefield==0.5.5
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
jellyfish==0.11.2
Levenshtein==0.27.1
lib50==3.0.11
Markdown==3.7
numpy==2.2.5
packaging==24.1
panda==0.3.1
pandas==2.2.3
pexpect==4.9.0
pillow==11.0.0
ptyprocess==0.7.0
pycparser==2.22
PyMySQL==1.1.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.0
reportlab==4.4.3
requests==2.32.3
setuptools==75.3.0
six==1.17.0
sqlparse==0.5.1
submit50==3.1.5
termcolor==1.1.0
tzdata==2025.2
urllib3==2.2.3
